Futureproof firefox rpath example
After a firefox upgrade, merely patching libxul.so was insufficient to
get apulse working on my machine.

In addition I use $LIB instead of lib so that the example can be used
unchanged for 32-bit applications on 64-bit hosts.

Also, since it uses find and file, it could theoretically be used to
patch an entire /usr tree safely.  Whether or not to recommend this in
the readme I leave up to future changes.